fixed one function and changed types

+++ types.h
@@ -1,8 +1,8 @@
 
-typedef double          Q_Type;
-typedef double          b_Type;                  //non metterlo mai unsigned
-typedef double          A_Type;
-typedef double          fx_Type;                //non metterlo mai unsigned
+typedef float          Q_Type;
+typedef int          b_Type;                  //non metterlo mai unsigned
+typedef int          A_Type;
+typedef float          fx_Type;                //non metterlo mai unsigned
 typedef unsigned char   dim_Type;
 typedef float           lambda_Type;            //choose only between floatand double
 typedef float           mu_Type;
